Hw7

+++ scrapper/src/main/java/edu/java/scrapper/controller/ScrapperController.java
@@ -6,8 +6,14 @@ import edu.java.scrapper.dto.scrapper.ScrapperPostRequest;
 import edu.java.scrapper.dto.scrapper.ScrapperPostResponse;
 import edu.java.scrapper.service.interfaces.ChatService;
 import edu.java.scrapper.service.interfaces.LinkService;
+import io.github.bucket4j.Bandwidth;
+import io.github.bucket4j.Bucket;
+import io.github.bucket4j.Refill;
+import jakarta.servlet.http.HttpServletRequest;
+import java.time.Duration;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.cache.caffeine.CaffeineCacheManager;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Component;
@@ -23,39 +29,71 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @Slf4j
 public class ScrapperController {
+
+    private static final String CACHE_NAME = "buckets";
+    private static final Long TOKENS_1 = 10L;
+    private static final Long TOKENS_2 = 5L;
+    private static final Long DURATION = 20L;
+
     private final ChatService chatService;
     private final LinkService linkService;
+    private final CaffeineCacheManager cacheManager;
+
+    @Autowired
+    private HttpServletRequest context;
 
     @Autowired
     public ScrapperController(
         ChatService chatService,
-        LinkService linkService
+        LinkService linkService,
+        CaffeineCacheManager cacheManager
     ) {
         this.chatService = chatService;
         this.linkService = linkService;
+        this.cacheManager = cacheManager;
     }
 
     @PostMapping(value = "/tg-chat/{id}")
     public ResponseEntity<Void> regChat(@PathVariable Long id) {
-        chatService.register(id, "default"); // stub for user name
-        return new ResponseEntity<>(HttpStatus.OK);
+        String ip = context.getRemoteAddr();
+        Bucket bucket = getBucket(ip);
+
+        if (bucket.tryConsume(1)) {
+            chatService.register(id, "default"); // stub for user name
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.TOO_MANY_REQUESTS);
     }
 
     @DeleteMapping(value = "/tg-chat/{id}")
     public ResponseEntity<Void> deleteChat(@PathVariable Long id) {
-        chatService.unregister(id);
-        return new ResponseEntity<>(HttpStatus.OK);
+        String ip = context.getRemoteAddr();
+        Bucket bucket = getBucket(ip);
+
+        if (bucket.tryConsume(1)) {
+            chatService.unregister(id);
+            return new ResponseEntity<>(HttpStatus.OK);
+        }
+
+        return new ResponseEntity<>(HttpStatus.TOO_MANY_REQUESTS);
     }
 
     @GetMapping(value = "/links", produces = "application/json")
     public ResponseEntity<ScrapperGetResponse> getAllLinks(
         @RequestHeader(name = "Tg-Chat-Id") Long id
     ) {
-        var links = linkService.listAll(id);
-        return new ResponseEntity<>(
-            new ScrapperGetResponse(links, links.size()),
-            HttpStatus.OK
-        );
+        String ip = context.getRemoteAddr();
+        Bucket bucket = getBucket(ip);
+
+        if (bucket.tryConsume(1)) {
+            var links = linkService.listAll(id);
+            return new ResponseEntity<>(
+                new ScrapperGetResponse(links, links.size()),
+                HttpStatus.OK
+            );
+        }
+
+        return new ResponseEntity<>(HttpStatus.TOO_MANY_REQUESTS);
     }
 
     @PostMapping(value = "/links", consumes = "application/json", produces = "application/json")
@@ -63,13 +101,20 @@ public class ScrapperController {
         @RequestHeader(name = "Tg-Chat-Id") Long id,
         @RequestBody ScrapperPostRequest request
     ) {
-        log.info("Adding link");
-        var link = request.link().url();
-        linkService.add(id, link.toString().hashCode(), link);
-        return new ResponseEntity<>(
-            new ScrapperPostResponse(id, link),
-            HttpStatus.OK
-        );
+        String ip = context.getRemoteAddr();
+        Bucket bucket = getBucket(ip);
+
+        if (bucket.tryConsume(1)) {
+            log.info("Adding link");
+            var link = request.link().url();
+            linkService.add(id, link.toString().hashCode(), link);
+            return new ResponseEntity<>(
+                new ScrapperPostResponse(id, link),
+                HttpStatus.OK
+            );
+        }
+
+        return new ResponseEntity<>(HttpStatus.TOO_MANY_REQUESTS);
     }
 
     @DeleteMapping(value = "/links", consumes = "application/json", produces = "application/json")
@@ -77,10 +122,37 @@ public class ScrapperController {
         @RequestHeader(name = "Tg-Chat-Id") Long id,
         @RequestBody ScrapperPostRequest link
     ) {
-        linkService.remove(id, link.link().url().toString().hashCode());
-        return new ResponseEntity<>(
-            new ScrapperDeleteResponse(id, link.link().url()),
-            HttpStatus.OK
-        );
+        String ip = context.getRemoteAddr();
+        Bucket bucket = getBucket(ip);
+
+        if (bucket.tryConsume(1)) {
+            linkService.remove(id, link.link().url().toString().hashCode());
+            return new ResponseEntity<>(
+                new ScrapperDeleteResponse(id, link.link().url()),
+                HttpStatus.OK
+            );
+        }
+
+        return new ResponseEntity<>(HttpStatus.TOO_MANY_REQUESTS);
+    }
+
+    private Bucket bucketSupplier() {
+        return Bucket.builder()
+            .addLimit(Bandwidth.classic(TOKENS_1, Refill.intervally(TOKENS_1, Duration.ofMinutes(1))))
+            .addLimit(Bandwidth.classic(TOKENS_2, Refill.intervally(TOKENS_2, Duration.ofSeconds(DURATION))))
+            .build();
+    }
+
+    private Bucket getBucket(String ip) {
+        var wrapper = cacheManager.getCache(CACHE_NAME).get(ip);
+        Bucket bucket;
+        if (wrapper != null) {
+            bucket = (Bucket) wrapper.get();
+        } else {
+            bucket = bucketSupplier();
+            cacheManager.getCache(CACHE_NAME).put(ip, bucket);
+        }
+
+        return bucket;
     }
 }
